feat: 권한 문제 수정 및 과제제안 필드 변경

+++ src/main/java/com/scg/stop/proposal/controller/ProposalController.java
@@ -1,6 +1,8 @@
 package com.scg.stop.proposal.controller;
 
 import com.scg.stop.auth.annotation.AuthUser;
+import com.scg.stop.global.exception.BadRequestException;
+import com.scg.stop.global.exception.ExceptionCode;
 import com.scg.stop.proposal.domain.request.CreateProposalRequest;
 import com.scg.stop.proposal.domain.request.ProposalReplyRequest;
 import com.scg.stop.proposal.domain.response.ProposalDetailResponse;
@@ -9,6 +11,7 @@ import com.scg.stop.proposal.domain.response.ProposalResponse;
 import com.scg.stop.proposal.service.ProposalService;
 import com.scg.stop.user.domain.AccessType;
 import com.scg.stop.user.domain.User;
+import com.scg.stop.user.domain.UserType;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -37,14 +40,14 @@ public class ProposalController {
     public ResponseEntity<Page<ProposalResponse>> getProposals(@AuthUser(accessType = {AccessType.COMPANY, AccessType.ADMIN}) User user,
                                                                @RequestParam(value = "title", required = false) String title,
                                                                @PageableDefault(page = 0, size = 10) Pageable pageable) {
-        Page<ProposalResponse> proposalResponse = proposalService.getProposalList(title, pageable);
+        Page<ProposalResponse> proposalResponse = proposalService.getProposalList(title, pageable, user);
         return ResponseEntity.status(HttpStatus.OK).body(proposalResponse);
     }
 
     @GetMapping("/{proposalId}")
     public ResponseEntity<ProposalDetailResponse> getProposal(@AuthUser(accessType = {AccessType.COMPANY, AccessType.ADMIN}) User user,
                                                               @PathVariable("proposalId") Long proposalId) {
-        ProposalDetailResponse proposalDetailResponse = proposalService.getProposalDetail(proposalId);
+        ProposalDetailResponse proposalDetailResponse = proposalService.getProposalDetail(proposalId, user);
         return ResponseEntity.status(HttpStatus.OK).body(proposalDetailResponse);
     }
 
@@ -60,14 +63,14 @@ public class ProposalController {
                                                                  @PathVariable("proposalId") Long proposalId,
                                                                  @RequestBody @Valid CreateProposalRequest proposalUpdateRequest) {
         ProposalDetailResponse proposalDetailResponse = proposalService.updateProposal(proposalId,
-                proposalUpdateRequest);
+                proposalUpdateRequest, user);
         return ResponseEntity.status(HttpStatus.OK).body(proposalDetailResponse);
     }
 
     @DeleteMapping("/{proposalId}")
     public ResponseEntity<Void> deleteProposal(@AuthUser(accessType = {AccessType.ADMIN, AccessType.COMPANY}) User user,
                                            @PathVariable("proposalId") Long proposalId) {
-        proposalService.deleteProposal(proposalId);
+        proposalService.deleteProposal(proposalId, user);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
@@ -75,6 +78,7 @@ public class ProposalController {
     public ResponseEntity<ProposalReplyResponse> createProposalReply(@AuthUser(accessType = {AccessType.ADMIN}) User user,
                                                                      @PathVariable("proposalId") Long proposalId,
                                                                      @RequestBody @Valid ProposalReplyRequest proposalReplyCreateRequest) {
+        if (user.getUserType() != UserType.ADMIN) throw new BadRequestException(ExceptionCode.NOT_AUTHORIZED);
         ProposalReplyResponse proposalReplyResponse = proposalService.createProposalReply(proposalId, proposalReplyCreateRequest);
         return ResponseEntity.status(HttpStatus.CREATED).body(proposalReplyResponse);
     }
@@ -85,12 +89,14 @@ public class ProposalController {
                                                                      @PathVariable("proposalReplyId") Long proposalReplyId,
                                                                      @RequestBody @Valid ProposalReplyRequest proposalReplyUpdateRequest) {
 
+        if (user.getUserType() != UserType.ADMIN) throw new BadRequestException(ExceptionCode.NOT_AUTHORIZED);
         ProposalReplyResponse proposalReplyResponse = proposalService.updateProposalReply(proposalReplyId, proposalReplyUpdateRequest);
         return ResponseEntity.status(HttpStatus.OK).body(proposalReplyResponse);
     }
     @DeleteMapping("/{proposalId}/reply/{proposalReplyId}")
     public ResponseEntity<Void> deleteProposalReply(@AuthUser(accessType = {AccessType.ADMIN}) User user,
                                            @PathVariable("proposalId") Long proposalId, @PathVariable("proposalReplyId") Long proposalReplyId) {
+        if (user.getUserType() != UserType.ADMIN) throw new BadRequestException(ExceptionCode.NOT_AUTHORIZED);
         proposalService.deleteProposalReply(proposalReplyId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
@@ -99,7 +105,7 @@ public class ProposalController {
     public ResponseEntity<ProposalReplyResponse> getProposalReplies(@AuthUser(accessType = {AccessType.ADMIN}) User user,
                                                                     @PathVariable("proposalId") Long proposalId) {
 //        List<ProposalReplyResponse> proposalReplies = proposalService.getProposalReplies(proposalId);
-        ProposalReplyResponse proposalReply = proposalService.getProposalReply(proposalId);
+        ProposalReplyResponse proposalReply = proposalService.getProposalReply(proposalId, user);
         return ResponseEntity.status(HttpStatus.OK).body(proposalReply);
     }
 }
